+++ jest.config.js
@@ -135,7 +135,6 @@ module.exports = {
   setupFilesAfterEnv: ['<rootDir>/src/jest-setup.ts'],
 
   // A list of paths to snapshot serializer modules Jest should use for snapshot testing
-  snapshotSerializers: ['enzyme-to-json/serializer'],
 
   // The test environment that will be used for testing
   testEnvironment: 'jest-environment-jsdom',