+++ src/components/animated-progress.test.tsx
@@ -3,7 +3,7 @@ import { ThemeProvider } from 'styled-components';
 
 import { DeckContext, DeckContextType } from './deck/deck';
 import defaultTheme from '../theme/default-theme';
-import Progress from './progress';
+import AnimatedProgress from './animated-progress';
 import { DeepPartial } from '../types/deep-partial';
 import { render } from '@testing-library/react';
 
@@ -24,28 +24,15 @@ const mountWithContext = (
   return render(tree, { wrapper: WrappingThemeProvider });
 };
 
-describe('<Progress />', () => {
+describe('<AnimatedProgress />', () => {
   it('should render the right amount of circles', () => {
-    const { queryAllByTestId } = mountWithContext(<Progress />, {
+    const { queryAllByTestId } = mountWithContext(<AnimatedProgress />, {
       slideCount: 5,
       activeView: {
         slideIndex: 0
       }
     });
 
-    expect(queryAllByTestId('Progress Circle')).toHaveLength(5);
-  });
-
-  it('should render the right amount of circles with the current circle in the active state', () => {
-    const { queryAllByTestId } = mountWithContext(<Progress />, {
-      slideCount: 5,
-      activeView: {
-        slideIndex: 4
-      }
-    });
-
-    expect(queryAllByTestId('Progress Circle')[4]).toHaveStyle({
-      background: '#fff'
-    });
+    expect(queryAllByTestId('animated-progress-circle')).toHaveLength(5);
   });
 });