+++ src/components/table.test.tsx
@@ -1,23 +1,21 @@
-import { FC, ReactElement } from 'react';
-import Enzyme, { mount } from 'enzyme';
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+import { FC, PropsWithChildren, ReactElement } from 'react';
 import { ThemeProvider } from 'styled-components';
 
 import defaultTheme from '../theme/default-theme';
 import { Table, TableRow, TableCell, TableBody, TableHeader } from './table';
-
-Enzyme.configure({ adapter: new Adapter() });
+import { render } from '@testing-library/react';
 
 const mountWithTheme = (tree: ReactElement) => {
-  const WrappingThemeProvider: FC = (props) => (
+  const WrappingThemeProvider = (props: PropsWithChildren) => (
     <ThemeProvider theme={defaultTheme}>{props.children}</ThemeProvider>
   );
-  return mount(tree, { wrappingComponent: WrappingThemeProvider });
+
+  return render(tree, { wrapper: WrappingThemeProvider });
 };
 
 describe('<Table />', () => {
   it('should render a <table> with a <tr> for each row and a <td> with text for each cell', () => {
-    const wrapper = mountWithTheme(
+    const { container } = mountWithTheme(
       <Table>
         <TableBody>
           <TableRow>
@@ -31,20 +29,29 @@ describe('<Table />', () => {
         </TableBody>
       </Table>
     );
-    expect(wrapper).toMatchSnapshot();
+
+    expect(container.querySelector('table')).toBeDefined();
+
+    const row1 = container.querySelectorAll('tr')[0],
+      row2 = container.querySelectorAll('tr')[1];
+    expect(row1.querySelectorAll('td')).toHaveLength(2);
+    expect(row2.querySelectorAll('td')).toHaveLength(2);
   });
 
   it('should render a <table> with bold text for the header row', () => {
-    const wrapper = mountWithTheme(
+    const { container } = mountWithTheme(
       <Table>
         <TableHeader>
           <TableRow>
-            <TableCell>Row 1</TableCell>
-            <TableCell>Row 1</TableCell>
+            <TableCell>Row 1, Col 1</TableCell>
+            <TableCell>Row 1, Col 2</TableCell>
           </TableRow>
         </TableHeader>
       </Table>
     );
-    expect(wrapper).toMatchSnapshot();
+
+    expect(container.querySelector('thead')).toHaveStyle({
+      fontWeight: 'bold'
+    });
   });
 });